+++ tablests/stringify.ts
