+++ vitest.config.ts
@@ -1,21 +1,21 @@
-import { defineConfig } from "vitest/config";
-import { resolve } from "path";
-import dts from "vite-plugin-dts";
+import { defineConfig } from 'vitest/config';
+import { resolve } from 'path';
+import dts from 'vite-plugin-dts';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  build: {
-    lib: {
-      entry: resolve(__dirname, "csvts/csvts.ts"),
-      name: "csvts",
-      // the proper extensions will be added
-      fileName: "csvts",
+    build: {
+        lib: {
+            entry: resolve(__dirname, 'tablests/tables.ts'),
+            name: 'tablests',
+            // the proper extensions will be added
+            fileName: 'tablests',
+        },
     },
-  },
-  resolve: {
-    alias: {
-      "@csvts": resolve(__dirname, "./csvts"),
+    resolve: {
+        alias: {
+            '@tablests': resolve(__dirname, './tablests'),
+        },
     },
-  },
-  plugins: [dts()],
+    plugins: [dts()],
 });
